quick fix : Redundant type cast

+++ src/main/java/tconstruct/client/BlockSkinRenderHelper.java
@@ -188,7 +188,7 @@ public class BlockSkinRenderHelper {
             render.colorRedTopRight *= f6;
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
-            render.renderFaceYNeg(block, (double) xPos, (double) yPos, (double) zPos, block.getIcon(0, metadata));
+            render.renderFaceYNeg(block, xPos, yPos, zPos, block.getIcon(0, metadata));
             flag = true;
         }
 
@@ -308,7 +308,7 @@ public class BlockSkinRenderHelper {
             render.colorRedTopRight *= f6;
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
-            render.renderFaceYPos(block, (double) xPos, (double) yPos, (double) zPos, block.getIcon(1, metadata));
+            render.renderFaceYPos(block, xPos, yPos, zPos, block.getIcon(1, metadata));
             flag = true;
         }
 
@@ -442,7 +442,7 @@ public class BlockSkinRenderHelper {
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
             icon = block.getIcon(2, metadata);
-            render.renderFaceZNeg(block, (double) xPos, (double) yPos, (double) zPos, icon);
+            render.renderFaceZNeg(block, xPos, yPos, zPos, icon);
 
             flag = true;
         }
@@ -575,7 +575,7 @@ public class BlockSkinRenderHelper {
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
             icon = block.getIcon(3, metadata);
-            render.renderFaceZPos(block, (double) xPos, (double) yPos, (double) zPos, icon);
+            render.renderFaceZPos(block, xPos, yPos, zPos, icon);
 
             flag = true;
         }
@@ -708,7 +708,7 @@ public class BlockSkinRenderHelper {
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
             icon = block.getIcon(4, metadata);
-            render.renderFaceXNeg(block, (double) xPos, (double) yPos, (double) zPos, icon);
+            render.renderFaceXNeg(block, xPos, yPos, zPos, icon);
 
             flag = true;
         }
@@ -841,7 +841,7 @@ public class BlockSkinRenderHelper {
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
             icon = block.getIcon(5, metadata);
-            render.renderFaceXPos(block, (double) xPos, (double) yPos, (double) zPos, icon);
+            render.renderFaceXPos(block, xPos, yPos, zPos, icon);
 
             flag = true;
         }
@@ -891,7 +891,7 @@ public class BlockSkinRenderHelper {
                     render.renderMinY > 0.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos, yPos - 1, zPos));
             tessellator.setColorOpaque_F(f10, f13, f16);
-            render.renderFaceYNeg(block, (double) xPos, (double) yPos, (double) zPos, block.getIcon(0, metadata));
+            render.renderFaceYNeg(block, xPos, yPos, zPos, block.getIcon(0, metadata));
             flag = true;
         }
 
@@ -900,7 +900,7 @@ public class BlockSkinRenderHelper {
                     render.renderMaxY < 1.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos, yPos + 1, zPos));
             tessellator.setColorOpaque_F(f7, f8, f9);
-            render.renderFaceYPos(block, (double) xPos, (double) yPos, (double) zPos, block.getIcon(1, metadata));
+            render.renderFaceYPos(block, xPos, yPos, zPos, block.getIcon(1, metadata));
             flag = true;
         }
 
@@ -911,7 +911,7 @@ public class BlockSkinRenderHelper {
                     render.renderMinZ > 0.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos, yPos, zPos - 1));
             tessellator.setColorOpaque_F(f11, f14, f17);
-            render.renderFaceZNeg(block, (double) xPos, (double) yPos, (double) zPos, block.getIcon(2, metadata));
+            render.renderFaceZNeg(block, xPos, yPos, zPos, block.getIcon(2, metadata));
 
             flag = true;
         }
@@ -921,7 +921,7 @@ public class BlockSkinRenderHelper {
                     render.renderMaxZ < 1.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos, yPos, zPos + 1));
             tessellator.setColorOpaque_F(f11, f14, f17);
-            render.renderFaceZPos(block, (double) xPos, (double) yPos, (double) zPos, block.getIcon(3, metadata));
+            render.renderFaceZPos(block, xPos, yPos, zPos, block.getIcon(3, metadata));
 
             flag = true;
         }
@@ -931,7 +931,7 @@ public class BlockSkinRenderHelper {
                     render.renderMinX > 0.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos - 1, yPos, zPos));
             tessellator.setColorOpaque_F(f12, f15, f18);
-            render.renderFaceXNeg(block, (double) xPos, (double) yPos, (double) zPos, block.getIcon(4, metadata));
+            render.renderFaceXNeg(block, xPos, yPos, zPos, block.getIcon(4, metadata));
 
             flag = true;
         }
@@ -941,7 +941,7 @@ public class BlockSkinRenderHelper {
                     render.renderMaxX < 1.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos + 1, yPos, zPos));
             tessellator.setColorOpaque_F(f12, f15, f18);
-            render.renderFaceXPos(block, (double) xPos, (double) yPos, (double) zPos, block.getIcon(5, metadata));
+            render.renderFaceXPos(block, xPos, yPos, zPos, block.getIcon(5, metadata));
 
             flag = true;
         }
@@ -1176,7 +1176,7 @@ public class BlockSkinRenderHelper {
             render.colorRedTopRight *= f6;
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
-            render.renderFaceYNeg(block, (double) xPos, (double) yPos, (double) zPos, stillIcon);
+            render.renderFaceYNeg(block, xPos, yPos, zPos, stillIcon);
             flag = true;
         }
 
@@ -1296,7 +1296,7 @@ public class BlockSkinRenderHelper {
             render.colorRedTopRight *= f6;
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
-            render.renderFaceYPos(block, (double) xPos, (double) yPos, (double) zPos, stillIcon);
+            render.renderFaceYPos(block, xPos, yPos, zPos, stillIcon);
             flag = true;
         }
 
@@ -1427,7 +1427,7 @@ public class BlockSkinRenderHelper {
             render.colorRedTopRight *= f6;
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
-            render.renderFaceZNeg(block, (double) xPos, (double) yPos, (double) zPos, flowingIcon);
+            render.renderFaceZNeg(block, xPos, yPos, zPos, flowingIcon);
 
             flag = true;
         }
@@ -1559,7 +1559,7 @@ public class BlockSkinRenderHelper {
             render.colorRedTopRight *= f6;
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
-            render.renderFaceZPos(block, (double) xPos, (double) yPos, (double) zPos, flowingIcon);
+            render.renderFaceZPos(block, xPos, yPos, zPos, flowingIcon);
 
             flag = true;
         }
@@ -1691,7 +1691,7 @@ public class BlockSkinRenderHelper {
             render.colorRedTopRight *= f6;
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
-            render.renderFaceXNeg(block, (double) xPos, (double) yPos, (double) zPos, flowingIcon);
+            render.renderFaceXNeg(block, xPos, yPos, zPos, flowingIcon);
 
             flag = true;
         }
@@ -1823,7 +1823,7 @@ public class BlockSkinRenderHelper {
             render.colorRedTopRight *= f6;
             render.colorGreenTopRight *= f6;
             render.colorBlueTopRight *= f6;
-            render.renderFaceXPos(block, (double) xPos, (double) yPos, (double) zPos, flowingIcon);
+            render.renderFaceXPos(block, xPos, yPos, zPos, flowingIcon);
 
             flag = true;
         }
@@ -1874,7 +1874,7 @@ public class BlockSkinRenderHelper {
                     render.renderMinY > 0.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos, yPos - 1, zPos));
             tessellator.setColorOpaque_F(f10, f13, f16);
-            render.renderFaceYNeg(block, (double) xPos, (double) yPos, (double) zPos, stillIcon);
+            render.renderFaceYNeg(block, xPos, yPos, zPos, stillIcon);
             flag = true;
         }
 
@@ -1883,7 +1883,7 @@ public class BlockSkinRenderHelper {
                     render.renderMaxY < 1.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos, yPos + 1, zPos));
             tessellator.setColorOpaque_F(f7, f8, f9);
-            render.renderFaceYPos(block, (double) xPos, (double) yPos, (double) zPos, stillIcon);
+            render.renderFaceYPos(block, xPos, yPos, zPos, stillIcon);
             flag = true;
         }
 
@@ -1894,7 +1894,7 @@ public class BlockSkinRenderHelper {
                     render.renderMinZ > 0.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos, yPos, zPos - 1));
             tessellator.setColorOpaque_F(f11, f14, f17);
-            render.renderFaceZNeg(block, (double) xPos, (double) yPos, (double) zPos, flowingIcon);
+            render.renderFaceZNeg(block, xPos, yPos, zPos, flowingIcon);
 
             flag = true;
         }
@@ -1904,7 +1904,7 @@ public class BlockSkinRenderHelper {
                     render.renderMaxZ < 1.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos, yPos, zPos + 1));
             tessellator.setColorOpaque_F(f11, f14, f17);
-            render.renderFaceZPos(block, (double) xPos, (double) yPos, (double) zPos, flowingIcon);
+            render.renderFaceZPos(block, xPos, yPos, zPos, flowingIcon);
 
             flag = true;
         }
@@ -1914,7 +1914,7 @@ public class BlockSkinRenderHelper {
                     render.renderMinX > 0.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos - 1, yPos, zPos));
             tessellator.setColorOpaque_F(f12, f15, f18);
-            render.renderFaceXNeg(block, (double) xPos, (double) yPos, (double) zPos, flowingIcon);
+            render.renderFaceXNeg(block, xPos, yPos, zPos, flowingIcon);
 
             flag = true;
         }
@@ -1924,7 +1924,7 @@ public class BlockSkinRenderHelper {
                     render.renderMaxX < 1.0D ? l
                             : block.getMixedBrightnessForBlock(render.blockAccess, xPos + 1, yPos, zPos));
             tessellator.setColorOpaque_F(f12, f15, f18);
-            render.renderFaceXPos(block, (double) xPos, (double) yPos, (double) zPos, flowingIcon);
+            render.renderFaceXPos(block, xPos, yPos, zPos, flowingIcon);
 
             flag = true;
         }
